+++ src/cli/.env
@@ -1,0 +1,2 @@
+RSS_name1=value1
+RSS_name2=value2
